+++ packages/react/src/index.ts
@@ -1,2 +1,6 @@
-export { useLdhopQuery } from './useLdhopQuery.js'
+export {
+  configureQueryKey,
+  DEFAULT_QUERY_KEY,
+  useLdhopQuery,
+} from './useLdhopQuery.js'
 export { useLDhopQuery } from './useLDhopQueryLegacy.js'